refactor: extract publishing submission flow
